requirements.txt
pandas
numpy
seaborn
matplotlib
scikit-learn
scipy
transformers
datasets
accelerate
bitsandbytes
openai
peft
pytorch
pytorch-cuda
sentencepiece
tokenizers
adapters